+++ CHANGELOG.md
@@ -7,8 +7,6 @@ _Next planned release.  Currently under development._
 * Added back the `catsoopdb` format (last seen in
     [version 4.0.1](https://gitlab.com/adqm/cat-soop/blob/master/CHANGELOG.md#version-401)),
     with improvements to prevent collisions and a few bugfixes.
-* Added [pure-Python scrypt module](https://github.com/ricmoo/pyscrypt) to
-    the distribution.
 * Added [Beautiful Soup](https://www.crummy.com/software/BeautifulSoup/) to the
     distribution.
 
@@ -17,6 +15,8 @@ _Next planned release.  Currently under development._
 * Renamed `gb.py` -> `base_context.py` to more accurately reflect its usage.
 * Moved `main.template` and `base.css` to `old.template` and `old.css`,
     respectively, to make room for new (more modern) style.
+* Modified the `login` authentication type to use Python's built-in
+    implementation of PBKDF2
 
 **Deprecated:**